remove dtype argument

+++ eval.py
@@ -16,7 +16,7 @@ if __name__ == "__main__":
     from utils import Evaluator, TimeStamp, set_gen_length, save_result
 
     args.kv_type = "retain"  # RetainCache enables efficient evaluation across multiple compression ratios with a single prefilling.
-    model = ModelKVzip(args.model, dtype=args.dtype, kv_type=args.kv_type)
+    model = ModelKVzip(args.model, kv_type=args.kv_type)
 
     dataset = load_dataset_all(args.data, model.tokenizer)  # list of data
     dataset = DataWrapper(args.data, dataset, model)
