+++ test.py
@@ -6,7 +6,7 @@ from model import ModelKVzip
 if __name__ == "__main__":
     from args import args
 
-    model = ModelKVzip(args.model, dtype=args.dtype, kv_type=args.kv_type)
+    model = ModelKVzip(args.model, kv_type=args.kv_type)
 
     dataset = load_dataset_all(args.data, model.tokenizer)  # list of data
     dataset = DataWrapper(args.data, dataset, model)
